Restructure into sophisticated lib/ based architecture
User instruction: "the structure itself needs attention its bad right now search
online web search research for the best structure most used most sophisticated
structure possible and something that gnome allows research then implement"

Research findings (verified via web search of popular GNOME extensions):
- Just Perfection (1.9M+ downloads): uses lib/ with subdirectories
- Blur My Shell: uses src/ (but has Makefile build step)
- Rounded Window Corners: uses named subdirs at root
- gjs.guide confirms: no folder name restrictions, modules can be organized freely

Conclusion: lib/ is the most widely-used convention for GNOME extensions
WITHOUT a build step. src/ implies compilation/transformation.

File restructuring:
- spotlightPopup.js    → lib/ui/spotlightPopup.js
- popup/popupBackdrop.js  → lib/ui/popupBackdrop.js
- popup/popupPositioner.js → lib/ui/popupPositioner.js
- services/core/keybinding.js → lib/core/keybinding.js
- Empty popup/ and services/ directories removed

Imports updated:
- extension.js: ./lib/ui/spotlightPopup.js, ./lib/core/keybinding.js
- lib/ui/spotlightPopup.js: sibling imports (./popupBackdrop.js, ./popupPositioner.js)

Documentation updated:
- README.md: Rewrote file layout table and architecture description
- AGENTS.md: Updated file layout section
- .github/labeler.yml: Updated path globs (lib/ui/**, lib/core/keybinding.js)

Skills updated:
- skills/extension-best-practices/SKILL.md: Added "directory structure
  convention" section documenting the lib/ pattern with rationale and
  reference to Just Perfection as the leading example.

All extension logic code untouched. All CI passes.
