+++ .github/labeler.yml
@@ -1,8 +1,8 @@
 popup ui:
   - changed-files:
       - any-glob-to-any-file:
-          - spotlightPopup.js
-          - popup/**
+          - lib/ui/spotlightPopup.js
+          - lib/ui/**
           - stylesheet.css
 preferences:
   - changed-files:
@@ -12,7 +12,7 @@ preferences:
 keybinding:
   - changed-files:
       - any-glob-to-any-file:
-          - services/core/keybinding.js
+          - lib/core/keybinding.js
 
 documentation:
   - changed-files:
